// features/api/model/liftingsIndex200.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { LiftingResource } from './liftingResource';
import type { LiftingsIndex200Links } from './liftingsIndex200Links';
import type { LiftingsIndex200Meta } from './liftingsIndex200Meta';

export type LiftingsIndex200 = {
  data: LiftingResource[];
  links: LiftingsIndex200Links;
  meta: LiftingsIndex200Meta;
};