// features/api/model/liftingsIndex200Meta.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { LiftingsIndex200MetaLinksItem } from './liftingsIndex200MetaLinksItem';

export type LiftingsIndex200Meta = {
  /** @minimum 1 */
  current_page: number;
  /**
   * @minimum 1
   * @nullable
   */
  from: number | null;
  /** @minimum 1 */
  last_page: number;
  /** Generated paginator links. */
  links: LiftingsIndex200MetaLinksItem[];
  /**
   * Base path for paginator generated URLs.
   * @nullable
   */
  path: string | null;
  /**
   * Number of items shown per page.
   * @minimum 0
   */
  per_page: number;
  /**
   * Number of the last item in the slice.
   * @minimum 1
   * @nullable
   */
  to: number | null;
  /**
   * Total number of items being paginated.
   * @minimum 0
   */
  total: number;
};